flask_app_code/requirements.txt
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.1
flask==3.0.3
gunicorn==22.0.0
importlib-metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
tomli==2.0.1
werkzeug==3.0.3
zipp==3.19.2
